% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{menuItem_cwee}
\alias{menuItem_cwee}
\title{menuItem_cwee}
\usage{
menuItem_cwee(text, ..., icon = NULL, badgeLabel = NULL,
  badgeColor = "green", tabName = NULL, href = NULL, newtab = TRUE,
  disable = FALSE)
}
\description{
menuItem_cwee
}

