% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotLoadProfile.R
\name{plotFullTS_js_kW}
\alias{plotFullTS_js_kW}
\title{plotFullTS_js_kW}
\usage{
plotFullTS_js_kW(dat)
}
\arguments{
\item{dat}{data...}
}
\value{
\code{NULL}
}
\description{
plotFullTS_js_kW
}

