% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotLoadProfile.R
\name{plotFullTS_r}
\alias{plotFullTS_r}
\title{plotFullTS_r}
\usage{
plotFullTS_r(dat, inludeZero = TRUE)
}
\arguments{
\item{dat}{data...}

\item{inludeZero}{inludeZero...}
}
\value{
\code{NULL}
}
\description{
plotFullTS_r
}

