% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotHeatmap_8760.R
\name{plotHeatmap_8760}
\alias{plotHeatmap_8760}
\title{plotLoadProfile_ggplot}
\usage{
plotHeatmap_8760(dat, var, normVar, inclWeekends)
}
\arguments{
\item{dat}{data...}
}
\value{
\code{NULL}
}
\description{
plotLoadProfile_ggplot
}

