% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotLoadProfile.R
\name{plotLoadProfile_dygraph}
\alias{plotLoadProfile_dygraph}
\title{plotLoadProfile_dygraph}
\usage{
plotLoadProfile_dygraph(dat, var, type)
}
\arguments{
\item{dat}{data...}

\item{type}{type...}
}
\value{
\code{NULL}
}
\description{
plotLoadProfile_dygraph
}

