% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotLoadProfile.R
\name{plotLoadProfile_ggplot}
\alias{plotLoadProfile_ggplot}
\title{plotLoadProfile_ggplot}
\usage{
plotLoadProfile_ggplot(dat, var, type, color, labelMax = FALSE, yRng)
}
\arguments{
\item{dat}{data...}

\item{type}{type...}

\item{color}{color...}

\item{labelMax}{labelMax...}
}
\value{
\code{NULL}
}
\description{
plotLoadProfile_ggplot
}

